Add workflow to detect errors in the antora files
Currently errors in the antora build are detected earliest at the stage
of the final build in the website. Actually, that's to late for a
developer to respond.
The error on the page build should occur as early as possible during the
PR.

Therefor, this PR commits a new workflow that just builds the website.
Using the loglevel definition during the build leads to the build
returning an exit status code in case any errors occur.