fix: Propose anemic domain models
The reason for the anemic domain models lies in two things.
First of all are the domain models returned to the adapter.
But business logic should not be moved outside.
An additional mapping should be avoided.
Furthermore in large scaled applications anemic domain models are much
easier to handle, because the business logic is not spread across use
cases, domain models and services.
Making the applicaiton easier to understand